requirements.txt
numpy
pandas
scikit-learn
sentence-transformers
xgboost
matplotlib
seaborn
cupy-cuda12x